Moved inistream to its own file
